pyproject.toml
[tool.poetry]
name = "medical-imaging-review-agent"
version = "0.1.0"
description = "To write later"
authors = ["Fahim <shahriarahmadf@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
openai = "^1.0.0"
langchain = "^0.3.23"
pandas = "^2.0"
numpy = "^1.25"
scikit-learn = "^1.3"
langchain-openai = "^0.3.14"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
